refactor: create execution/merge_queue/ subpackage
Move merge_queue.py → merge_queue/__init__.py and 3 _merge_queue_*.py
modules into execution/merge_queue/. Update all internal imports.
